♻️ :  Separate model files
+ Lint some codes
