+++ Targets/MorningBear/Sources/Home/Article/Article.swift
@@ -1,0 +1,15 @@
+//
+//  Article.swift
+//  MorningBear
+//
+//  Created by 이영빈 on 2022/12/27.
+//  Copyright © 2022 com.dache. All rights reserved.
+//
+
+import UIKit
+
+struct Article {
+    let image: UIImage
+    let title: String
+    let desc: String
+}
