+++ Targets/MorningBear/Sources/Home/Badge/Badge.swift
@@ -1,0 +1,16 @@
+//
+//  Badge.swift
+//  MorningBear
+//
+//  Created by 이영빈 on 2022/12/27.
+//  Copyright © 2022 com.dache. All rights reserved.
+//
+
+import UIKit
+
+/// 나의 배지 정보
+struct Badge {
+    let image: UIImage
+    let title: String
+    let desc: String
+}
