+++ Targets/MorningBear/Sources/Home/Common/HomeSection.swift
@@ -18,27 +18,4 @@ enum HomeSection {
     case badges([Badge])
     /// 읽을만한 글
     case articles([Article])
-    
-    // FIXME: 아래 모델들 다 나중에 별개 파일로 분리시킬 예정
-    struct State {
-        let nickname: String
-    }
-    
-    struct RecentMorning {
-        let image: UIImage
-        let title: String
-        let desc: String
-    }
-    
-    struct Badge {
-        let image: UIImage
-        let title: String
-        let desc: String
-    }
-    
-    struct Article {
-        let image: UIImage
-        let title: String
-        let desc: String
-    }
 }
