+++ Targets/MorningBear/Sources/Home/State/State.swift
@@ -1,0 +1,13 @@
+//
+//  State.swift
+//  MorningBear
+//
+//  Created by 이영빈 on 2022/12/27.
+//  Copyright © 2022 com.dache. All rights reserved.
+//
+
+import Foundation
+
+struct State {
+    let nickname: String
+}
